Added guide access restriction
