+++ client/src/components/common/NavBar.tsx
@@ -54,6 +54,7 @@ const NavBar: React.FC<Props> = ({}) => {
             role={[Role.Professor]}
             children={<Button className="navbutton" href="/my-pending-case-studies">Mes études de cas</Button>}
         ></UnlockAccess>
+        <Button className="navbutton" href="/guide">Guides</Button>
         <span>
           <Button
               id="basic-button"
