+++ client/src/components/guidePage/GuidePage.scss
@@ -7,4 +7,15 @@
     display: flex;
     flex-direction: column;
 }
+
+.login-prompt {
+    margin: 0.5em;
+    width: fit-content;
+}
+
+.login-prompt-text {
+    font-size: 1em;
+    font-weight: bold;
+    text-decoration: underline;
+}
