+++ client/src/components/guidePage/GuidePage.tsx
@@ -2,12 +2,20 @@ import React from "react";
 import PdfAccordion from "./PdfAccordion";
 import "./GuidePage.scss";
 import NavBar from "../common/NavBar";
+import { Button, Typography } from "@mui/material";
+import Cookies from "js-cookie";
+import LoginPopup, { LoginPopupRef } from "./../connection/LoginPopup";
 
 function GuidePage() {
-  const numberOfPdfFiles = 10;
-  const pdfFiles = [];
-  const pdfTitles = [
-    "Définition d'un cas",
+  const loginPopupRef = React.useRef<LoginPopupRef | null>(null);
+
+  const numberOfUnrestrictedPdfFiles = 1;
+  const numberOfRestrictedPdfFiles = 9;
+  const totalNumberOfPdfFiles = numberOfUnrestrictedPdfFiles + numberOfRestrictedPdfFiles;
+  const unrestrictedPdfFiles = [];
+  const restrictedPdfFiles = [];
+  const unrestrictedPdfTitles = ["Définition d'un cas"];
+  const restrictedPdfTitles = [
     "Processus de création d'un cas",
     "Écriture d'un cas",
     "Écriture d'une note pédagogique d'un cas",
@@ -19,24 +27,55 @@ function GuidePage() {
     "Conseils pour animer un cas",
   ];
 
-  for (let i = 1; i <= numberOfPdfFiles; i++) {
-    pdfFiles.push(require(`../../pdfs/fiche-${i}.pdf`));
+  for (let i = 1; i <= numberOfUnrestrictedPdfFiles; i++) {
+    unrestrictedPdfFiles.push(require(`../../pdfs/fiche-${i}.pdf`));
   }
+  for (let i = numberOfUnrestrictedPdfFiles + 1; i <= totalNumberOfPdfFiles; i++) {
+    restrictedPdfFiles.push(require(`../../pdfs/fiche-${i}.pdf`));
+  }
+
+  const openPopup = () => {
+    if (loginPopupRef.current) {
+      loginPopupRef.current.setPopupOpen();
+    }
+  };
 
   return (
     <>
-    <NavBar></NavBar>
-    <div className="guide-page">
-      <h1>Ressources pédagogiques</h1>
-      <h2>Comment rédiger et animer une étude de cas ?</h2>
-      <div className="accordions-container">
-        {pdfFiles.map((pdfFile, index) => (
-          <PdfAccordion key={index} pdfFile={pdfFile} index={index + 1} title={pdfTitles[index]} />
-        ))}
+      <NavBar></NavBar>
+      <div className="guide-page">
+        <h1>Ressources pédagogiques</h1>
+        <h2>Comment rédiger et animer une étude de cas ?</h2>
+        <div className="accordions-container">
+          {unrestrictedPdfFiles.map((pdfFile, index) => (
+            <PdfAccordion
+              key={index}
+              pdfFile={pdfFile}
+              index={index + 1}
+              title={unrestrictedPdfTitles[index]}
+              disabled={false}
+            />
+          ))}
+          {!Cookies.get("logged_in") && (
+            <Button className="login-prompt" onClick={openPopup}>
+              <Typography className="login-prompt-text">
+                Connectez-vous pour consulter le reste des fiches!
+              </Typography>
+            </Button>
+          )}
+          {restrictedPdfFiles.map((pdfFile, index) => (
+            <PdfAccordion
+              key={index}
+              pdfFile={pdfFile}
+              index={index + 1 + numberOfUnrestrictedPdfFiles}
+              title={restrictedPdfTitles[index]}
+              disabled={!Cookies.get("logged_in")}
+            />
+          ))}
+        </div>
+        <LoginPopup ref={loginPopupRef}/>
       </div>
-    </div>
     </>
-    
   );
 }
